Refactor commands into a single open command
