+++ bin/cli.js
@@ -2,18 +2,6 @@
 
 const yargs = require("yargs");
 const checkKit = require("check-kit");
-const open = require("open");
-
-const locations = {
-  website: "http://romajs.org",
-  twitter: "https://twitter.com/roma_js",
-  facebook: "https://www.facebook.com/romajs.org",
-  youtube: "https://www.youtube.com/channel/UCFm8OPi5USbFybw9SaTLxeA",
-  slack: "https://romajs.slack.com/",
-  "slack-signup": "https://romajs.herokuapp.com/",
-  github: "https://github.com/Roma-JS",
-  meetup: "http://www.meetup.com/RomaJS/",
-};
 
 (async () => {
   const { current, name, latest, updateAvailable } = await checkKit.check({
@@ -23,16 +11,8 @@ const locations = {
   if (updateAvailable) {
     console.log(`New version of ${name} available! ${current} -> ${latest}`);
   } else {
-    for (const [name, url] of Object.entries(locations)) {
-      yargs.command({
-        command: name,
-        desc: `Open RomaJS ${name}`,
-        handler: async () => {
-          await open(url);
-        },
-      });
-    }
     yargs
+      .commandDir("../cmds")
       .wrap(yargs.terminalWidth())
       .showHelpOnFail(true)
       .demandCommand()
